Test failure case
* Before, we did not test this and it was hitting a url that does not
  exist in our app! Silly us.
