+++ spec/rails_helper.rb
@@ -9,6 +9,8 @@ require 'capybara/rspec'
 
 Dir[Rails.root.join('spec/support/**/*.rb')].each { |f| require f }
 
+include OmniAuthHelper
+
 ActiveRecord::Migration.maintain_test_schema!
 
 RSpec.configure do |config|
