+++ spec/support/helpers/omni_auth_helper.rb
@@ -1,0 +1,21 @@
+module OmniAuthHelper
+  def current_user_uid
+    '12345'
+  end
+
+  def mock_github(opts = { })
+    OmniAuth.config.mock_auth[:github] = github_auth_hash(opts)
+  end
+
+  def github_auth_hash(opts = { })
+    OmniAuth::AuthHash.new({
+      provider: 'github',
+      uid: current_user_uid,
+      info: {
+        name: 'Doris Doogooder',
+        email: 'test@example.com',
+        nickname: 'github_username'
+      }
+    }.merge(opts))
+  end
+end
